Removing Privoxy post
Got an odd email saying Privoxy website links to a scam site. Quick
investigation showed its hosting is sponsored by a binary options site
but otherwise did not see anything odd.

But regardless the post is so out of date I thought it would be safest
just to remove it.